@@ -9,41 +9,66 @@
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write ."
+    "format": "prettier --plugin-search-dir . --write .",
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^2.1.1",
-    "@sveltejs/kit": "^1.20.4",
-    "@types/gapi.client.drive-v3": "^0.0.4",
-    "@typescript-eslint/eslint-plugin": "^5.45.0",
-    "@typescript-eslint/parser": "^5.45.0",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.28.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte": "^2.30.0",
-    "flowbite-svelte": "^0.44.5",
-    "flowbite-svelte-icons": "^0.4.2",
+    "@sveltejs/adapter-auto": "^4.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/forms": "^0.5.10",
+    "@tailwindcss/typography": "^0.5.19",
+    "@tailwindcss/vite": "^4.1.17",
+    "@testing-library/svelte": "^5.2.7",
+    "@types/gapi.client.drive-v3": "^0.0.5",
+    "@types/google.accounts": "^0.0.15",
+    "@types/google.picker": "^0.0.49",
+    "@types/pica": "^9.0.5",
+    "@typescript-eslint/eslint-plugin": "^8.26.0",
+    "@typescript-eslint/parser": "^8.26.0",
+    "@vitest/coverage-v8": "^3.0.7",
+    "autoprefixer": "^10.4.20",
+    "caniuse-lite": "^1.0.30001747",
+    "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.0.2",
+    "eslint-plugin-svelte": "^3.0.2",
+    "flowbite": "^4.0.1",
+    "flowbite-svelte": "^1.28.1",
+    "flowbite-svelte-icons": "^3.0.0",
+    "jsdom": "^26.0.0",
     "postcss": "^8.4.24",
-    "postcss-load-config": "^4.0.1",
-    "prettier": "^2.8.0",
-    "prettier-plugin-svelte": "^2.10.1",
+    "postcss-load-config": "^6.0.1",
+    "prettier": "^3.1.0",
+    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-tailwindcss": "^0.7.1",
     "sass": "^1.64.2",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
-    "tailwindcss": "^3.3.2",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.3.4",
+    "tailwindcss": "^4.1.17",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^4.4.2"
+    "typescript": "^5.5.0",
+    "vite": "^6.2.0",
+    "vitest": "^3.0.7"
   },
   "type": "module",
   "dependencies": {
     "@types/gapi": "^0.0.47",
-    "@types/google.accounts": "^0.0.14",
-    "@types/google.picker": "^0.0.42",
-    "@vercel/analytics": "^1.1.0",
-    "@zip.js/zip.js": "^2.7.20",
+    "@vercel/analytics": "^1.5.0",
+    "@zip.js/zip.js": "^2.7.57",
+    "chart.js": "^4.5.1",
     "dexie": "^4.0.1-alpha.25",
+    "megajs": "^1.3.9",
+    "page-flip": "^2.0.7",
     "panzoom": "^9.4.3",
-    "svelte-easy-crop": "^2.0.1"
+    "pica": "^9.0.1",
+    "svelte-easy-crop": "^4.0.0",
+    "webdav": "^5.8.0"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "esbuild",
+      "@tailwindcss/oxide"
+    ]
   }
 }
