@@ -1,23 +1,27 @@
 <script lang="ts">
   import { toggleFullScreen, zoomFitToScreen } from '$lib/panzoom';
-  import { SpeedDial, SpeedDialButton } from 'flowbite-svelte';
   import { settings } from '$lib/settings';
   import {
     ArrowLeftOutline,
     ArrowRightOutline,
     CompressOutline,
     ImageOutline,
-    ZoomOutOutline
+    ZoomOutOutline,
+    PlusOutline
   } from 'flowbite-svelte-icons';
   import { imageToWebp, showCropper, updateLastCard } from '$lib/anki-connect';
   import { promptConfirmation } from '$lib/util';
 
-  export let left: (_e: any, ingoreTimeOut?: boolean) => void;
-  export let right: (_e: any, ingoreTimeOut?: boolean) => void;
-  export let src1: File;
-  export let src2: File | undefined;
+  interface Props {
+    left: (_e: any, ingoreTimeOut?: boolean) => void;
+    right: (_e: any, ingoreTimeOut?: boolean) => void;
+    src1: File | undefined;
+    src2: File | undefined;
+  }
 
-  let open = false;
+  let { left, right, src1, src2 }: Props = $props();
+
+  let open = $state(false);
 
   function handleZoom() {
     zoomFitToScreen();
@@ -47,37 +51,84 @@
     }
     open = false;
   }
+
+  function toggleMenu() {
+    open = !open;
+  }
 </script>
 
 {#if $settings.quickActions}
-  <SpeedDial
-    tooltip="none"
-    trigger="click"
-    defaultClass="absolute end-3 bottom-3 z-50"
-    color="transparent"
-    bind:open
-  >
-    {#if $settings.ankiConnectSettings.enabled}
-      <SpeedDialButton name={src2 ? '1' : undefined} on:click={() => onUpdateCard(src1)}>
-        <ImageOutline />
-      </SpeedDialButton>
+  <div class="fixed end-3 bottom-3 z-50 flex flex-col items-center">
+    <!-- Action buttons (shown when open) -->
+    {#if open}
+      <div class="mb-2 flex flex-col items-center gap-2">
+        {#if $settings.ankiConnectSettings.enabled}
+          <button
+            onclick={() => onUpdateCard(src1)}
+            class="relative flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+            aria-label="Add image to Anki"
+          >
+            <ImageOutline size="xl" />
+            {#if src2}
+              <span
+                class="absolute -top-1 -right-1 flex h-5 w-5 items-center justify-center rounded-full bg-primary-600 text-xs text-white"
+                >1</span
+              >
+            {/if}
+          </button>
+        {/if}
+        {#if $settings.ankiConnectSettings.enabled && src2}
+          <button
+            onclick={() => onUpdateCard(src2)}
+            class="relative flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+            aria-label="Add image 2 to Anki"
+          >
+            <ImageOutline size="xl" />
+            <span
+              class="absolute -top-1 -right-1 flex h-5 w-5 items-center justify-center rounded-full bg-primary-600 text-xs text-white"
+              >2</span
+            >
+          </button>
+        {/if}
+        <button
+          onclick={toggleFullScreen}
+          class="flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+          aria-label="Toggle fullscreen"
+        >
+          <CompressOutline size="xl" />
+        </button>
+        <button
+          onclick={handleZoom}
+          class="flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+          aria-label="Zoom to fit"
+        >
+          <ZoomOutOutline size="xl" />
+        </button>
+        <button
+          onclick={handleRight}
+          class="flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+          aria-label="Next page"
+        >
+          <ArrowRightOutline size="xl" />
+        </button>
+        <button
+          onclick={handleLeft}
+          class="flex h-12 w-12 items-center justify-center rounded-full bg-gray-700 text-gray-300 shadow-lg hover:bg-gray-600 focus:outline-none dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600"
+          aria-label="Previous page"
+        >
+          <ArrowLeftOutline size="xl" />
+        </button>
+      </div>
     {/if}
-    {#if $settings.ankiConnectSettings.enabled && src2}
-      <SpeedDialButton name="2" on:click={() => onUpdateCard(src2)}>
-        <ImageOutline />
-      </SpeedDialButton>
-    {/if}
-    <SpeedDialButton on:click={toggleFullScreen}>
-      <CompressOutline />
-    </SpeedDialButton>
-    <SpeedDialButton on:click={handleZoom}>
-      <ZoomOutOutline />
-    </SpeedDialButton>
-    <SpeedDialButton on:click={handleRight}>
-      <ArrowRightOutline />
-    </SpeedDialButton>
-    <SpeedDialButton on:click={handleLeft}>
-      <ArrowLeftOutline />
-    </SpeedDialButton>
-  </SpeedDial>
+
+    <!-- Main toggle button -->
+    <button
+      onclick={toggleMenu}
+      class="flex h-12 w-12 items-center justify-center rounded-full text-white focus:ring-2 focus:ring-blue-500 focus:outline-none"
+      aria-label="Quick actions menu"
+      style="transition: transform 0.3s ease; transform: rotate({open ? 45 : 0}deg);"
+    >
+      <PlusOutline size="xl" />
+    </button>
+  </div>
 {/if}
