@@ -1,27 +1,39 @@
 <script lang="ts">
+  import { preventDefault } from 'svelte/legacy';
+
   import {
+    BUILT_IN_PROFILES,
     copyProfile,
     createProfile,
     deleteProfile,
     profiles,
-    renameProfile
+    renameProfile,
+    type BuiltInProfile
   } from '$lib/settings';
   import { promptConfirmation, showSnackbar } from '$lib/util';
-  import { Listgroup, ListgroupItem, Modal, Input } from 'flowbite-svelte';
+  import { Input, Listgroup, ListgroupItem, Modal } from 'flowbite-svelte';
   import {
     CirclePlusSolid,
-    CopySolid,
     EditOutline,
+    FileCopySolid,
     TrashBinSolid,
     UserEditSolid
   } from 'flowbite-svelte-icons';
-  import type { ListGroupItemType } from 'flowbite-svelte/dist/types';
 
-  export let open = false;
+  interface Props {
+    open?: boolean;
+  }
 
-  $: items = Object.keys($profiles);
+  let { open = $bindable(false) }: Props = $props();
 
-  let newProfile: string;
+  let items = $derived(Object.keys($profiles));
+
+  let newProfile: string = $state('');
+
+  // Helper to check if a profile is built-in
+  function isBuiltIn(profileName: string): boolean {
+    return BUILT_IN_PROFILES.includes(profileName as BuiltInProfile);
+  }
 
   function onSubmit() {
     if (!newProfile) {
@@ -38,26 +50,39 @@
     newProfile = '';
   }
 
-  function onCopy(item: string | ListGroupItemType) {
+  function onCopy(item: string) {
     let newCopy = `${item} copy`;
 
     while (items.includes(newCopy)) {
       newCopy += ` copy`;
     }
 
-    copyProfile(item as string, newCopy);
+    copyProfile(item, newCopy);
   }
 
-  function onDelete(item: string | ListGroupItemType) {
+  function onDelete(item: string) {
+    if (isBuiltIn(item)) {
+      showSnackbar('Cannot delete built-in profile');
+      return;
+    }
+
     promptConfirmation(`Are you sure you would like to delete the [${item}] profile?`, () => {
-      deleteProfile(item as string);
+      const success = deleteProfile(item);
+      if (success) {
+        showSnackbar('Profile deleted');
+      }
     });
   }
 
-  let profileToEdit: string | ListGroupItemType;
-  let newName: string | ListGroupItemType;
+  let profileToEdit: string = $state('');
+  let newName: string = $state('');
 
-  function onEditClicked(item: string | ListGroupItemType) {
+  function onEditClicked(item: string) {
+    if (isBuiltIn(item)) {
+      showSnackbar('Cannot rename built-in profile');
+      return;
+    }
+
     if (profileToEdit) {
       profileToEdit = '';
     } else {
@@ -67,12 +92,16 @@
   }
 
   function onEdit() {
-    if (items.includes(newName as string)) {
+    if (items.includes(newName)) {
       showSnackbar('Profile already exists');
       return;
     }
 
-    renameProfile(profileToEdit as string, newName as string);
+    const success = renameProfile(profileToEdit, newName);
+    if (success) {
+      profileToEdit = '';
+      showSnackbar('Profile renamed');
+    }
   }
 
   function onInputClick(this: any) {
@@ -81,40 +110,49 @@
 </script>
 
 <Modal size="xs" bind:open outsideclose>
-  <Listgroup {items} let:item>
-    <ListgroupItem class="flex flex-row justify-between gap-6">
-      <div class="flex-1">
-        {#if profileToEdit === item}
-          <form on:submit|preventDefault={onEdit}>
-            <Input size="sm" bind:value={newName} autofocus on:click={onInputClick}>
-              <EditOutline
-                slot="right"
-                size="sm"
-                on:click={onEdit}
-                class="hover:text-primary-700"
-              />
-            </Input>
-          </form>
-        {:else}
-          <p class="line-clamp-1">{item}</p>
-        {/if}
-      </div>
-      <div class="flex flex-row gap-2 items-center">
-        <CopySolid size="sm" class="hover:text-primary-700" on:click={() => onCopy(item)} />
-        {#if item !== 'Default'}
-          <UserEditSolid
-            size="sm"
-            class="hover:text-primary-700"
-            on:click={() => onEditClicked(item)}
-          />
-          <TrashBinSolid size="sm" class="hover:text-primary-700" on:click={() => onDelete(item)} />
-        {/if}
-      </div>
-    </ListgroupItem>
+  <Listgroup {items}>
+    {#snippet children(itemData)}
+      {@const item = String(itemData)}
+      <ListgroupItem class="flex flex-row justify-between gap-6">
+        <div class="flex flex-1 items-center gap-2">
+          {#if profileToEdit === item}
+            <form onsubmit={preventDefault(onEdit)}>
+              <Input size="sm" bind:value={newName} autofocus onclick={onInputClick}>
+                {#snippet right()}
+                  <EditOutline size="sm" onclick={onEdit} class="hover:text-primary-700" />
+                {/snippet}
+              </Input>
+            </form>
+          {:else}
+            <p class="line-clamp-1">{item}</p>
+            {#if isBuiltIn(item)}
+              <span class="rounded-full bg-blue-500 px-2 py-0.5 text-xs text-white">Built-in</span>
+            {/if}
+          {/if}
+        </div>
+        <div class="flex flex-row items-center gap-2">
+          <FileCopySolid size="sm" class="hover:text-primary-700" onclick={() => onCopy(item)} />
+          {#if !isBuiltIn(item)}
+            <UserEditSolid
+              size="sm"
+              class="hover:text-primary-700"
+              onclick={() => onEditClicked(item)}
+            />
+            <TrashBinSolid
+              size="sm"
+              class="hover:text-primary-700"
+              onclick={() => onDelete(item)}
+            />
+          {/if}
+        </div>
+      </ListgroupItem>
+    {/snippet}
   </Listgroup>
-  <form on:submit|preventDefault={onSubmit}>
+  <form onsubmit={preventDefault(onSubmit)}>
     <Input type="text" placeholder="New profile..." bind:value={newProfile}>
-      <CirclePlusSolid slot="right" class="hover:text-primary-700" on:click={onSubmit} />
+      {#snippet right()}
+        <CirclePlusSolid class="hover:text-primary-700" onclick={onSubmit} />
+      {/snippet}
     </Input>
   </form>
 </Modal>
