@@ -1,16 +1,28 @@
 <script lang="ts">
-  import { changeProfile, currentProfile, profiles } from '$lib/settings';
+  import {
+    changeProfile,
+    currentProfile,
+    profiles,
+    profilesWithTrash,
+    migrateProfiles
+  } from '$lib/settings';
   import { AccordionItem, Button, Select } from 'flowbite-svelte';
   import ManageProfilesModal from './ManageProfilesModal.svelte';
   import { showSnackbar } from '$lib/util';
 
-  export let onClose: () => void;
+  interface Props {
+    onClose: () => void;
+  }
 
-  $: items = Object.keys($profiles).map((id) => {
-    return { value: id, name: id };
-  });
+  let { onClose }: Props = $props();
 
-  let profile = $currentProfile;
+  let items = $derived(
+    Object.keys($profiles).map((id) => {
+      return { value: id, name: id };
+    })
+  );
+
+  let profile = $state($currentProfile);
 
   function onChange() {
     changeProfile(profile);
@@ -26,17 +38,20 @@
     showSnackbar('Profiles exported');
   }
 
-  let files: FileList;
+  let files: FileList | undefined = $state(undefined);
   function importProfile() {
+    if (!files) return;
     const [file] = files;
     const reader = new FileReader();
 
     reader.onloadend = () => {
       const imported = JSON.parse(reader.result?.toString() || '');
-      profiles.update((prev) => {
+      // Migrate imported profiles to ensure all fields exist with defaults
+      const migrated = migrateProfiles(imported);
+      profilesWithTrash.update((prev) => {
         return {
           ...prev,
-          ...imported
+          ...migrated
         };
       });
       onClose();
@@ -48,27 +63,32 @@
     }
   }
 
-  let manageModalOpen = false;
+  let manageModalOpen = $state(false);
 </script>
 
 <ManageProfilesModal bind:open={manageModalOpen} />
 
 <AccordionItem>
-  <span slot="header">Profile</span>
+  {#snippet header()}Profile{/snippet}
   <div class="flex flex-col gap-5">
     <div class="flex flex-col gap-2">
-      <Select {items} bind:value={profile} on:change={onChange} placeholder="Select profile ..." />
-      <Button size="sm" outline color="dark" on:click={() => (manageModalOpen = true)}
+      <Select {items} bind:value={profile} onchange={onChange} placeholder="Select profile ..." />
+      <Button size="sm" outline color="dark" onclick={() => (manageModalOpen = true)}
         >Manage profiles</Button
       >
     </div>
     <hr class="border-gray-100 opacity-10" />
     <div class="flex flex-col gap-2">
-      <input class="border border-slate-700 rounded-lg" type="file" accept=".json" bind:files />
-      <Button on:click={importProfile} disabled={!files} size="sm" outline color="blue"
+      <input
+        class="rounded-lg border border-slate-700 text-gray-900 file:mr-4 file:rounded-md file:border-0 file:bg-gray-100 file:px-4 file:py-2 file:text-sm file:font-semibold file:text-gray-700 hover:file:bg-gray-200 dark:text-white dark:file:bg-gray-700 dark:file:text-gray-200 dark:hover:file:bg-gray-600"
+        type="file"
+        accept=".json"
+        bind:files
+      />
+      <Button onclick={importProfile} disabled={!files} size="sm" outline color="blue"
         >Import profiles</Button
       >
-      <Button on:click={exportProfiles} size="sm" color="light">Export profiles</Button>
+      <Button onclick={exportProfiles} size="sm" color="light">Export profiles</Button>
     </div>
   </div>
 </AccordionItem>
