@@ -1,41 +1,73 @@
 <script lang="ts">
-  import { page } from '$app/stores';
-  import { zoomDefault } from '$lib/panzoom';
   import {
+    effectiveVolumeSettings,
     updateProgress,
     updateVolumeSetting,
     volumes,
-    volumeSettings,
-    type VolumeSettingsKey
+    type VolumeSettingsKey,
+    type PageViewMode
   } from '$lib/settings';
-  import { AccordionItem, Helper, Toggle } from 'flowbite-svelte';
+  import { zoomDefault } from '$lib/panzoom';
+  import { AccordionItem, Helper, Toggle, Label, Select } from 'flowbite-svelte';
+  import { routeParams } from '$lib/util/hash-router';
 
-  const volumeId = $page.params.volume;
+  const volumeId = $routeParams.volume!;
 
-  $: settings = $volumeSettings[$page.params.volume];
+  let settings = $derived($effectiveVolumeSettings[volumeId]);
 
-  $: toggles = [
+  let toggles = $derived([
     { key: 'rightToLeft', text: 'Right to left', value: settings.rightToLeft },
-    { key: 'singlePageView', text: 'Single page view', value: settings.singlePageView },
-    { key: 'hasCover', text: 'First page is cover', value: settings.hasCover }
-  ] as { key: VolumeSettingsKey; text: string; value: any }[];
+    { key: 'hasCover', text: 'First page is cover', value: settings.hasCover, shortcut: 'C' }
+  ] as { key: VolumeSettingsKey; text: string; value: any; shortcut?: string }[]);
+
+  const pageViewModes: { value: PageViewMode; name: string }[] = [
+    { value: 'single', name: 'Single page' },
+    { value: 'dual', name: 'Dual page' },
+    { value: 'auto', name: 'Auto (detect orientation & spreads)' }
+  ];
 
   function onChange(key: VolumeSettingsKey, value: any) {
-    updateVolumeSetting(volumeId, key, !value);
     if (key === 'hasCover') {
+      updateVolumeSetting(volumeId, key, !value);
       const pageClamped = Math.max($volumes[volumeId].progress - 1, 1);
       updateProgress(volumeId, pageClamped);
       zoomDefault();
+    } else {
+      updateVolumeSetting(volumeId, key, !value);
     }
+  }
+
+  function onPageViewModeChange(event: Event) {
+    const target = event.target as HTMLSelectElement;
+    updateVolumeSetting(volumeId, 'singlePageView', target.value as PageViewMode);
+    zoomDefault();
   }
 </script>
 
 <AccordionItem open>
-  <span slot="header">Volume settings</span>
+  {#snippet header()}Volume settings{/snippet}
   <div class="flex flex-col gap-5">
     <Helper>These settings only apply to this volume</Helper>
-    {#each toggles as { key, text, value }}
-      <Toggle size="small" checked={value} on:change={() => onChange(key, value)}>{text}</Toggle>
+    <div>
+      <Label for="page-view-mode" class="mb-2 text-gray-900 dark:text-white">
+        Page view mode
+        <span class="ml-2 text-xs text-gray-500 dark:text-gray-400">(P)</span>
+      </Label>
+      <Select
+        id="page-view-mode"
+        size="sm"
+        items={pageViewModes}
+        bind:value={settings.singlePageView}
+        onchange={onPageViewModeChange}
+      />
+    </div>
+    {#each toggles as { key, text, value, shortcut }}
+      <Toggle size="small" checked={value} onchange={() => onChange(key, value)}>
+        {text}
+        {#if shortcut}
+          <span class="ml-2 text-xs text-gray-500 dark:text-gray-400">({shortcut})</span>
+        {/if}
+      </Toggle>
     {/each}
   </div>
 </AccordionItem>
