@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { goto } from '$app/navigation';
   import { page } from '$app/stores';
   import Loader from '$lib/components/Loader.svelte';
   import { getItems, processFiles } from '$lib/upload';
   import { promptConfirmation, showSnackbar } from '$lib/util';
+  import { nav } from '$lib/util/navigation';
   import { P, Progressbar } from 'flowbite-svelte';
   import { onMount } from 'svelte';
   export const BASE_URL = $page.url.searchParams.get('source') || 'https://mokuro.moe/manga';
@@ -12,14 +12,14 @@
   const volume = $page.url.searchParams.get('volume');
   const url = `${BASE_URL}/${manga}/${volume}`;
 
-  let message = 'Loading...';
+  let message = $state('Loading...');
 
   let files: File[] = [];
 
-  let completed = 0;
-  let max = 0;
+  let completed = $state(0);
+  let max = $state(0);
 
-  $: progress = Math.floor((completed / max) * 100).toString();
+  let progress = $derived(Math.floor((completed / max) * 100).toString());
 
   async function onImport() {
     const mokuroRes = await fetch(url + '.mokuro', { cache: 'no-store' });
@@ -55,16 +55,15 @@
       completed++;
     }
     files.push(mokuroFile);
-    files = files;
     message = 'Adding to catalog...';
 
     processFiles(files).then(() => {
-      goto('/', { replaceState: true });
+      nav.toCatalog({ replaceState: true });
     });
   }
 
   function onCancel() {
-    goto('/', { replaceState: true });
+    nav.toCatalog({ replaceState: true });
   }
 
   onMount(() => {
