@@ -1,13 +1,6 @@
-const tailwindcss = require('tailwindcss');
-const autoprefixer = require('autoprefixer');
-
+// Tailwind CSS v4 is handled by @tailwindcss/vite plugin
 const config = {
-  plugins: [
-    //Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-    tailwindcss(),
-    //But others, like autoprefixer, need to run after,
-    autoprefixer
-  ]
+  plugins: {}
 };
 
 module.exports = config;
