@@ -1,7 +1,7 @@
 import { showSnackbar } from '$lib/util';
 import { writable } from 'svelte/store';
 import { blobToBase64, imageResize } from '.';
-import { Settings } from '$lib/settings';
+import type { Settings } from '$lib/settings/settings';
 
 type CropperModal = {
   open: boolean;
@@ -33,15 +33,20 @@ function getRadianAngle(degreeValue: number) {
   return (degreeValue * Math.PI) / 180;
 }
 
-export type Pixels = { width: number; height: number; x: number; y: number }
+export type Pixels = { width: number; height: number; x: number; y: number };
 
-export async function getCroppedImg(imageSrc: string, pixelCrop: Pixels, settings: Settings, rotation = 0 ) {
+export async function getCroppedImg(
+  imageSrc: string,
+  pixelCrop: Pixels,
+  settings: Settings,
+  rotation = 0
+) {
   const image = await createImage(imageSrc);
   const canvas = new OffscreenCanvas(image.width, image.height);
   const ctx = canvas.getContext('2d');
 
   if (!ctx) {
-    showSnackbar('Error: crop failed')
+    showSnackbar('Error: crop failed');
     return;
   }
 
@@ -72,9 +77,17 @@ export async function getCroppedImg(imageSrc: string, pixelCrop: Pixels, setting
     Math.round(0 - safeArea / 2 + image.width * 0.5 - pixelCrop.x),
     Math.round(0 - safeArea / 2 + image.height * 0.5 - pixelCrop.y)
   );
-  
-  await imageResize(canvas, ctx, settings.ankiConnectSettings.widthField, settings.ankiConnectSettings.heightField);
-  const blob = await canvas.convertToBlob({ type: 'image/webp', quality: settings.ankiConnectSettings.qualityField });
 
-  return await blobToBase64(blob)
+  await imageResize(
+    canvas,
+    ctx,
+    settings.ankiConnectSettings.widthField,
+    settings.ankiConnectSettings.heightField
+  );
+  const blob = await canvas.convertToBlob({
+    type: 'image/webp',
+    quality: settings.ankiConnectSettings.qualityField
+  });
+
+  return await blobToBase64(blob);
 }
