@@ -1,24 +1,25 @@
-import { Settings, settings } from "$lib/settings";
-import { showSnackbar } from "$lib/util"
-import { get } from "svelte/store";
+import type { Settings } from '$lib/settings/settings';
+import { settings } from '$lib/settings';
+import { showSnackbar } from '$lib/util';
+import { get } from 'svelte/store';
 
-export * from './cropper'
+export * from './cropper';
 
 export async function ankiConnect(action: string, params: Record<string, any>) {
   try {
     const res = await fetch('http://127.0.0.1:8765', {
       method: 'POST',
       body: JSON.stringify({ action, params, version: 6 })
-    })
-    const json = await res.json()
+    });
+    const json = await res.json();
 
     if (json.error) {
-      throw new Error(json.error)
+      throw new Error(json.error);
     }
 
     return json.result;
   } catch (e: any) {
-    showSnackbar(`Error: ${e?.message ?? e}`)
+    showSnackbar(`Error: ${e?.message ?? e}`);
   }
 }
 
@@ -30,11 +31,11 @@ export async function getCardInfo(id: string) {
 export async function getLastCardId() {
   const notesToday = await ankiConnect('findNotes', { query: 'added:1' });
   const id = notesToday.sort().at(-1);
-  return id
+  return id;
 }
 
 export async function getLastCardInfo() {
-  const id = await getLastCardId()
+  const id = await getLastCardId();
   return await getCardInfo(id);
 }
 
@@ -53,57 +54,69 @@ export async function blobToBase64(blob: Blob) {
 export async function imageToWebp(source: File, settings: Settings) {
   const image = await createImageBitmap(source);
   const canvas = new OffscreenCanvas(image.width, image.height);
-  const context = canvas.getContext("2d");
+  const context = canvas.getContext('2d');
 
   if (context) {
     context.drawImage(image, 0, 0);
-    await imageResize(canvas, context, settings.ankiConnectSettings.widthField, settings.ankiConnectSettings.heightField);
-    const blob = await canvas.convertToBlob({ type: 'image/webp', quality: settings.ankiConnectSettings.qualityField });
+    await imageResize(
+      canvas,
+      context,
+      settings.ankiConnectSettings.widthField,
+      settings.ankiConnectSettings.heightField
+    );
+    const blob = await canvas.convertToBlob({
+      type: 'image/webp',
+      quality: settings.ankiConnectSettings.qualityField
+    });
     image.close();
 
     return await blobToBase64(blob);
   }
 }
 
-export async function imageResize(canvas: OffscreenCanvas,  ctx: OffscreenCanvasRenderingContext2D, maxWidth: number, maxHeight: number): Promise<OffscreenCanvas> {
+export async function imageResize(
+  canvas: OffscreenCanvas,
+  ctx: OffscreenCanvasRenderingContext2D,
+  maxWidth: number,
+  maxHeight: number
+): Promise<OffscreenCanvas> {
   return new Promise((resolve, reject) => {
     const widthRatio = maxWidth <= 0 ? 1 : maxWidth / canvas.width;
     const heightRatio = maxHeight <= 0 ? 1 : maxHeight / canvas.height;
     const ratio = Math.min(1, Math.min(widthRatio, heightRatio));
 
     if (ratio < 1) {
-        const newWidth = canvas.width * ratio;
-        const newHeight = canvas.height * ratio;
-        createImageBitmap(canvas, { resizeWidth: newWidth, resizeHeight: newHeight, resizeQuality: 'high' })
-            .then((sprite) => {
-                canvas.width = newWidth;
-                canvas.height = newHeight;
-                ctx.drawImage(sprite, 0, 0);
-                resolve(canvas);
-            })
-            .catch((e) => reject(e));
+      const newWidth = canvas.width * ratio;
+      const newHeight = canvas.height * ratio;
+      createImageBitmap(canvas, {
+        resizeWidth: newWidth,
+        resizeHeight: newHeight,
+        resizeQuality: 'high'
+      })
+        .then((sprite) => {
+          canvas.width = newWidth;
+          canvas.height = newHeight;
+          ctx.drawImage(sprite, 0, 0);
+          resolve(canvas);
+        })
+        .catch((e) => reject(e));
     } else {
-        resolve(canvas);
+      resolve(canvas);
     }
   });
 }
 
 export async function updateLastCard(imageData: string | null | undefined, sentence?: string) {
-  const {
-    overwriteImage,
-    enabled,
-    grabSentence,
-    pictureField,
-    sentenceField
-  } = get(settings).ankiConnectSettings;
+  const { overwriteImage, enabled, grabSentence, pictureField, sentenceField } =
+    get(settings).ankiConnectSettings;
 
   if (!enabled) {
-    return
+    return;
   }
 
-  showSnackbar('Updating last card...', 10000)
+  showSnackbar('Updating last card...', 10000);
 
-  const id = await getLastCardId()
+  const id = await getLastCardId();
 
   if (getCardAgeInMin(id) >= 5) {
     showSnackbar('Error: Card created over 5 minutes ago');
@@ -117,7 +130,7 @@ export async function updateLastCard(imageData: string | null | undefined, sente
   }
 
   if (overwriteImage) {
-    fields[pictureField] = ''
+    fields[pictureField] = '';
   }
 
   if (imageData) {
@@ -128,15 +141,17 @@ export async function updateLastCard(imageData: string | null | undefined, sente
         picture: {
           filename: `mokuro_${id}.webp`,
           data: imageData.split(';base64,')[1],
-          fields: [pictureField],
-        },
-      },
-    }).then(() => {
-      showSnackbar('Card updated!')
-    }).catch((e) => {
-      showSnackbar(e)
+          fields: [pictureField]
+        }
+      }
     })
+      .then(() => {
+        showSnackbar('Card updated!');
+      })
+      .catch((e) => {
+        showSnackbar(e);
+      });
   } else {
-    showSnackbar('Something went wrong')
+    showSnackbar('Something went wrong');
   }
 }
