@@ -4,11 +4,11 @@
 </script>
 
 <AccordionItem>
-  <span slot="header">Stats</span>
+  {#snippet header()}Stats{/snippet}
   <div>
-    <p>Completed volumes: {$totalStats?.completed || 0}</p>
-    <p>Pages read: {$totalStats?.pagesRead || 0}</p>
-    <p>Characters read: {$totalStats?.charsRead || 0}</p>
-    <p>Minutes read: {$totalStats?.minutesRead || 0}</p>
+    <p class="text-gray-900 dark:text-white">Completed volumes: {$totalStats?.completed || 0}</p>
+    <p class="text-gray-900 dark:text-white">Pages read: {$totalStats?.pagesRead || 0}</p>
+    <p class="text-gray-900 dark:text-white">Characters read: {$totalStats?.charsRead || 0}</p>
+    <p class="text-gray-900 dark:text-white">Minutes read: {$totalStats?.minutesRead || 0}</p>
   </div>
 </AccordionItem>
