@@ -3,15 +3,35 @@ import { writable } from 'svelte/store';
 
 export type MiscSettings = {
   galleryLayout: 'grid' | 'list';
-  gallerySorting: 'ASC' | 'DESC';
+  gallerySorting: 'ASC' | 'DESC' | 'SMART';
+  deviceRamGB: 4 | 8 | 16 | 32;
+  turboMode: boolean;
+  gdriveAutoReAuth: boolean;
 };
 
 export type MiscSettingsKey = keyof MiscSettings;
 
+// Detect device memory and set default RAM config
+function getDefaultRamSetting(): 4 | 8 | 16 | 32 {
+  if (browser) {
+    const deviceMemory = (navigator as any).deviceMemory;
+    if (deviceMemory !== undefined) {
+      // deviceMemory is capped at 8GB, but if it reports 8, assume 16GB+ is likely
+      if (deviceMemory >= 8) return 16;
+      if (deviceMemory >= 4) return 8;
+      if (deviceMemory >= 2) return 4;
+    }
+  }
+  return 4; // Conservative default
+}
+
 const defaultSettings: MiscSettings = {
   galleryLayout: 'grid',
-  gallerySorting: 'ASC',
-}
+  gallerySorting: 'SMART',
+  deviceRamGB: getDefaultRamSetting(),
+  turboMode: false, // Default to single-operation mode (patient users)
+  gdriveAutoReAuth: true // Keep users synced during long reading sessions
+};
 
 const stored = browser ? window.localStorage.getItem('miscSettings') : undefined;
 
