@@ -13,17 +13,53 @@ export type Page = {
   img_path: string;
 };
 
-export type MokuroData = {
-  version: string;
-  title: string;
-  title_uuid: string;
-  volume: string;
+export interface VolumeMetadata {
+  mokuro_version: string; // Empty string '' indicates image-only volume without OCR
+  series_title: string;
+  series_uuid: string;
+  volume_title: string;
+  volume_uuid: string;
+  page_count: number;
+  character_count: number;
+  // Cumulative character counts per page: [50, 120, 200] means page 3 has 200 total chars through it
+  page_char_counts: number[];
+
+  // Thumbnail (small ~10-20KB file) and dimensions for synchronous layout
+  thumbnail?: File;
+  thumbnail_width?: number;
+  thumbnail_height?: number;
+
+  // Placeholder fields for cloud-only volumes (not yet downloaded locally)
+  isPlaceholder?: boolean;
+
+  // Generic cloud storage fields (new multi-provider format)
+  cloudProvider?: 'google-drive' | 'mega' | 'webdav';
+  cloudFileId?: string;
+  cloudModifiedTime?: string;
+  cloudSize?: number;
+
+  // Legacy Drive-specific fields (kept for backward compatibility)
+  // When present without cloudProvider, assumed to be google-drive
+  driveFileId?: string;
+  driveModifiedTime?: string;
+  driveSize?: number;
+}
+
+// v3 table: volume_ocr
+export interface VolumeOCR {
   volume_uuid: string;
   pages: Page[];
-};
+}
 
-export type Volume = {
-  mokuroData: MokuroData;
-  volumeName: string;
+// v3 table: volume_files
+export interface VolumeFiles {
+  volume_uuid: string;
   files: Record<string, File>;
-};
+}
+
+// Combined view for API compatibility (assembled from volume_ocr + volume_files)
+export interface VolumeData {
+  volume_uuid: string;
+  pages: Page[];
+  files?: Record<string, File>;
+}
