@@ -2,5 +2,10 @@ export * from './snackbar';
 export * from './upload';
 export * from './misc';
 export * from './modals';
-export * from './zip'
-export * from './cloud'
+export * from './zip';
+export * from './cloud';
+export * from './sync-store';
+export * from './sync/providers/google-drive';
+export * from './activity-tracker';
+export * from './backup';
+export * from './platform';
